Move BigInt and BigDecimal to their own files
