+++ std/src/main/scala/cats/std/bigdecimal.scala
@@ -1,0 +1,8 @@
+package cats
+package std
+
+trait BigDecimalInstances {
+  implicit val bigDecimalShow: Show[BigDecimal] = new Show[BigDecimal] {
+    def show(f: BigDecimal): String = f.toString
+  }
+}
