+++ std/src/main/scala/cats/std/bigint.scala
@@ -1,0 +1,8 @@
+package cats
+package std
+
+trait BigIntInstances extends algebra.std.BigIntInstances {
+  implicit val bigIntShow: Show[BigInt] = new Show[BigInt] {
+    def show(f: BigInt): String = f.toString
+  }
+}
